requirements.txt
Flask==2.3.0
reportlab==4.0.4
groq==0.3.0
python-dotenv==1.0.0
httpx==0.24.1
Werkzeug>=2.3.0
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3
blinker>=1.6.2
gunicorn==21.2.0
